+++ CHANGELOG.md
@@ -13,8 +13,13 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-beta.189 (2019-09-26)
 
+#### :boom: Breaking Change
+
+* [Removed function as a value support for `in-view` directive](https://github.com/V4Fire/Client/pull/130)
+
 #### :house: Internal
 
+* [Added `in-view` directive value validator](https://github.com/V4Fire/Client/pull/130)
 * [b-slider `optionKey` property is now able to be a function](https://github.com/V4Fire/Client/pull/131)
 
 ## v3.0.0-beta.188 (2019-09-25)